requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
albumentations=1.0.3=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
blas=1.0=mkl
bottleneck=1.3.2=py38h2a96729_1
brotlipy=0.7.0=py38h294d835_1001
ca-certificates=2021.5.30=h5b45459_0
certifi=2021.5.30=py38haa244fe_0
cffi=1.14.6=py38hd8c33c5_0
chardet=4.0.0=py38haa244fe_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
cloudpickle=1.6.0=py_0
cryptography=3.4.7=py38hd7da0ea_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py_2
cytoolz=0.11.0=py38h294d835_3
dask-core=2021.7.1=pyhd8ed1ab_0
decorator=5.0.9=pyhd8ed1ab_0
enum34=1.1.10=py38h32f6830_2
freetype=2.10.4=hd328e21_0
fsspec=2021.7.0=pyhd8ed1ab_0
geos=3.8.0=h33f27b4_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
idna=3.1=pyhd3deb0d_0
imagecodecs-lite=2019.12.3=py38h347fdf6_3
imageio=2.9.0=py_0
imgaug=0.4.0=py_1
intel-openmp=2021.3.0=haa95532_3372
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.1=py38hbd9d945_1
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
locket=0.2.0=py_2
lz4-c=1.9.3=h2bbff1b_0
matplotlib-base=3.4.2=py38heae8d8c_0
mkl=2021.3.0=haa95532_524
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.0=py38h277e83a_2
mkl_random=1.2.2=py38hf11a4ad_0
networkx=2.3=py_0
ninja=1.10.2=h6d14046_1
numexpr=2.7.3=py38hb80d3ca_1
numpy=1.20.3=py38ha4e8547_0
numpy-base=1.20.3=py38hc2deb75_0
olefile=0.46=py_0
opencv=4.0.1=py38h2a7c758_0
openssl=1.1.1k=h8ffe710_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.0=py38hd77b12b_0
partd=1.2.0=pyhd8ed1ab_0
pathlib=1.0.1=py38haa244fe_4
pillow=8.3.1=py38h4fa10fc_0
pip=21.1.3=py38haa95532_0
pooch=1.4.0=pyhd8ed1ab_0
py-opencv=4.0.1=py38he44ac1e_0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38haa244fe_3
python=3.8.8=hdbf39b2_5
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.9.0=py3.8_cuda10.2_cudnn7_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py38h347fdf6_3
pyyaml=5.4.1=py38h294d835_0
requests=2.26.0=pyhd8ed1ab_0
scikit-image=0.18.2=py38h60cbd38_0
scipy=1.6.2=py38h66253e8_1
setuptools=52.0.0=py38haa95532_0
shapely=1.7.1=py38h210f175_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=h2bbff1b_0
tifffile=2019.7.26.2=py38_0
tk=8.6.10=he774522_0
toolz=0.11.1=py_0
torch-same-pad=0.1.0=pypi_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu102
tornado=6.1=py38h294d835_1
tqdm=4.61.2=pyhd3eb1b0_1
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
